requirements.txt
# --- Required Modules --- #
# discord.py: Interfaces with the Discord API.
# discord.ext.commands: Provides tools to define and manage bot commands.
# dotenv: Loads environment variables from a .env file.
# re: Handles text parsing via regular expressions.
# os: Accesses system-level environment variables.
# asyncio: Enables asynchronous operations.
# datetime: Manages time-related tasks like cooldowns.
# logging: Records events and errors for debugging.
# requests: Handles HTTP requests for features like web search.
# google.generativeai: Connects to the Gemini API for AI responses.
# urllib.parse: Encodes URLs, used in search functionality.
# fastapi: Async web framework for building APIs.
# uvicorn: ASGI server required to run FastAPI apps and handle HTTP requests.

discord.py
python-dotenv
requests
google-generativeai
fastapi
uvicorn